+++ apps/chaingraph-frontend/src/components/flow/nodes/ChaingraphNode/NodeBody.tsx
@@ -17,7 +17,7 @@ import type {
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 import type { INode } from '@badaitech/chaingraph-types'
-import { PortComponent } from 'components/flow/nodes/ChaingraphNode/PortComponent'
+import { PortComponent } from '@/components/flow/nodes/ChaingraphNode/PortComponent'
 import { memo, useMemo } from 'react'
 
 export interface NodeBodyProps {