+++ apps/chaingraph-frontend/src/providers/RootProvider.tsx
@@ -7,12 +7,12 @@
  */
 
 import type { PropsWithChildren } from 'react'
-import { TrpcProvider } from '@/api/trpc/provider'
-import { DndContextProvider, DndProvider } from '@/components/dnd'
-import { MenuPositionProvider } from '@/components/flow/components/context-menu'
-import { ThemeProvider } from '@/components/theme/ThemeProvider'
-import { ZoomProvider } from '@/providers/ZoomProvider'
 import { ReactFlowProvider } from '@xyflow/react'
+import { TrpcProvider } from '../api/trpc/provider'
+import { DndContextProvider, DndProvider } from '../components/dnd'
+import { MenuPositionProvider } from '../components/flow/components/context-menu'
+import { ThemeProvider } from '../components/theme/ThemeProvider'
+import { ZoomProvider } from './ZoomProvider'
 
 export function RootProvider({ children }: PropsWithChildren) {
   return (