+++ packages/chaingraph-nodes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/chaingraph-nodes
 
+## 0.1.5
+
+### Patch Changes
+
+- Fix and adjust chaingraph-frontend lib NPM package
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes